Se agrega el listar Clientes, falta implementar las funciones de modificar y eliminar clientes.

+++ MyMProyecto/FrmAltaCollar.cs
@@ -58,7 +58,7 @@ namespace MyMProyecto
             {
                 collar.Largo = (CodigoCollar)cboCodigoCollar.SelectedItem;
                 collar.Color = (ColorCollar)cboColor.SelectedItem;
-                collar.PrecioCompra = float.Parse(txtPrecio.Text);
+                collar.PrecioCompra = decimal.Parse(txtPrecio.Text);
                 collar.Cantidad = int.Parse(txtCantidad.Text);
 
                 negocio.agregar(collar);
